Fix the decoding regressions caused by the simplification of the matroska video source
Originally committed to SVN as r178.

+++ src/core/ffvideosource.h
@@ -108,7 +108,7 @@ private:
     CompressedStream *CS;
 	char ErrorMessage[256];
 	FFSourceResources<FFMS_VideoSource> Res;
-	int PacketNumber;
+	size_t PacketNumber;
 
 	void DecodeNextFrame();
 protected:
